+++ spec/models/story_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Story, type: :model do
   it {is_expected.to have_field(:title).of_type(String)}
   it {is_expected.to have_field(:summary).of_type(String)}
   it {is_expected.to have_field(:index).of_type(Integer)}
-  it {is_expected.to have_field(:points).of_type(Integer)}
+  it {is_expected.to have_field(:points).of_type(Float)}
   it {is_expected.to have_field(:status).of_type(String).with_default_value_of("open")}
   it {is_expected.to have_field(:created_at).of_type(DateTime)}
   it {is_expected.to have_field(:created_by).of_type(String)}
